clippy clean (rust 1.95): match guards, drop redundant cast + unused binding
Auto-applied collapsible_match guard conversions (format.rs companion check,
ChosenTypeEntersWithCounter, CantBeBlockedExceptBy) surfaced by the 1.95
toolchain; remove an i32->i32 cast in the bot face-burn pass and an unused
test binding. Whole-workspace clippy and the 9582-test suite are green.

+++ crabomination/src/server/bot.rs
@@ -1804,7 +1804,7 @@ fn pick_removal_ping(state: &GameState, seat: usize) -> Option<GameAction> {
                 continue;
             }
             for opp in 0..state.players.len() {
-                if state.same_team(opp, seat) || state.players[opp].life > *n as i32 {
+                if state.same_team(opp, seat) || state.players[opp].life > *n {
                     continue;
                 }
                 let action = GameAction::ActivateAbility {
